Recognize haskell filetype from shebang
The default implementation of Haskell (GHC) provides the runhaskell
binary for creating executable scripts in Haskell.  It is referenced as
`#!/usr/bin/env runhaskell` in the Haskell build scripts and documented
recommendations.

Recognize this (and any other haskell in the shebang, which seems
unlikely to cause conflicts) as the haskell filetype.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>